feat: enhance admin followup category management with Vue3/Composition API
- Migrate FollowupSettings to Composition API for better reactivity and performance
- Fix CategoryDialog form validation and data handling
- Add category editing functionality with proper state management
- Implement refreshAccordion method for dynamic category updates
- Separate knowledge routes into admin and public endpoints
- Enhance audit logging with DATA_CREATE and DATA_UPDATE actions
- Improve error handling and loading states throughout admin interface
- Fix model binding issues in settings toggles and form controls
