+++ backend/core/settings_schemas.py
@@ -12,6 +12,108 @@ logger = logging.getLogger(__name__)
 
 
 @dataclass
+class FollowUpSettings:
+    """Configuration for follow-up question generation."""
+
+    enabled: bool = True
+    service_type: str = "static"  # static, dynamic, contextual
+    max_questions: int = 1
+    relevance_threshold: float = 0.7
+    include_technical: bool = True
+    include_personal: bool = True
+    include_creative: bool = True
+    question_style: str = "conversational"  # formal, conversational, exploratory
+    custom_questions: Dict[str, List[str]] = field(default_factory=dict)  # Custom questions by category
+
+    def to_dict(self) -> dict:
+        """Convert to dictionary for JSON serialization."""
+        return asdict(self)
+
+    @classmethod
+    def from_dict(cls, data: dict) -> "FollowUpSettings":
+        """Create from dictionary with validation."""
+        # Validate and set defaults for missing keys
+        defaults = cls()
+        validated_data = {}
+
+        for key, default_value in asdict(defaults).items():
+            if key in data:
+                validated_data[key] = data[key]
+            else:
+                validated_data[key] = default_value
+
+        # Type validation
+        if not isinstance(validated_data["enabled"], bool):
+            validated_data["enabled"] = str(validated_data["enabled"]).lower() == "true"
+
+        if not isinstance(validated_data["max_questions"], int):
+            try:
+                validated_data["max_questions"] = int(validated_data["max_questions"])
+            except (ValueError, TypeError):
+                validated_data["max_questions"] = defaults.max_questions
+
+        # Ensure max_questions is within reasonable bounds
+        validated_data["max_questions"] = max(1, min(5, validated_data["max_questions"]))
+
+        if not isinstance(validated_data["relevance_threshold"], float):
+            try:
+                validated_data["relevance_threshold"] = float(validated_data["relevance_threshold"])
+            except (ValueError, TypeError):
+                validated_data["relevance_threshold"] = defaults.relevance_threshold
+
+        # Ensure relevance_threshold is within bounds
+        validated_data["relevance_threshold"] = max(0.1, min(1.0, validated_data["relevance_threshold"]))
+
+        # Validate service_type
+        valid_service_types = ["static", "dynamic", "contextual"]
+        if validated_data["service_type"] not in valid_service_types:
+            validated_data["service_type"] = defaults.service_type
+
+        # Validate question_style
+        valid_styles = ["formal", "conversational", "exploratory"]
+        if validated_data["question_style"] not in valid_styles:
+            validated_data["question_style"] = defaults.question_style
+
+        # Validate custom_questions
+        if "custom_questions" in validated_data and validated_data["custom_questions"]:
+            custom_questions = validated_data["custom_questions"]
+            if not isinstance(custom_questions, dict):
+                validated_data["custom_questions"] = {}
+            else:
+                # Validate categories and questions
+                valid_categories = ["technical", "personal", "creative"]
+                cleaned_questions = {}
+                for category, questions in custom_questions.items():
+                    if category in valid_categories and isinstance(questions, list):
+                        # Filter and validate individual questions
+                        valid_questions = []
+                        for q in questions:
+                            if isinstance(q, str) and q.strip() and len(q.strip()) <= 200:
+                                valid_questions.append(q.strip())
+                        if valid_questions:
+                            cleaned_questions[category] = valid_questions[:20]  # Max 20 questions per category
+                validated_data["custom_questions"] = cleaned_questions
+        else:
+            validated_data["custom_questions"] = {}
+
+        return cls(**validated_data)
+
+    def to_json(self) -> str:
+        """Convert to JSON string for database storage."""
+        return json.dumps(self.to_dict())
+
+    @classmethod
+    def from_json(cls, json_str: str) -> "FollowUpSettings":
+        """Create from JSON string with error handling."""
+        try:
+            data = json.loads(json_str)
+            return cls.from_dict(data)
+        except (json.JSONDecodeError, TypeError, ValueError) as e:
+            logger.warning(f"Failed to parse follow-up settings from JSON: {e}")
+            return cls()  # Return defaults on error
+
+
+@dataclass
 class ResponseSettings:
     """Configuration for response generation behavior."""
 
@@ -243,7 +345,6 @@ class SystemSettings:
     def from_dict(cls, data: dict) -> "SystemSettings":
         """Create from dictionary with validation."""
         # Import here to avoid circular import
-        from .config import FollowUpSettings
 
         followup_data = data.get("followup", {})
         response_data = data.get("response", {})
