+++ tests/test_followup_settings.py
@@ -111,7 +111,7 @@ class TestFollowUpSettings:
     def test_followup_settings_json_error_handling(self):
         """Test JSON parsing error handling."""
         # Invalid JSON should return defaults
-        incomplete_json = "{"  # Incomplete JSON: opening brace without closing brace
+        incomplete_json = "{"  # Incomplete JSON with unclosed brace
         settings = FollowUpSettings.from_json(incomplete_json)
 
         # Should be default settings
